% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ang_diff}
\alias{ang_diff}
\title{ang_diff}
\usage{
ang_diff(target, pressed)
}
\arguments{
\item{target}{numeric the first angle in degrees}

\item{pressed}{numeric the second angle in degrees}
}
\value{
numeric the difference in degrees
}
\description{
compute the closest angular difference between two angles. Thanks
to thanks to https://stackoverflow.com/a/7869457
}
