% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_dist}
\alias{calc_dist}
\title{calc_dist}
\usage{
calc_dist(center_x = 0, center_y = 0, point_x, point_y)
}
\arguments{
\item{center_x}{numeric the \code{x} coordinate of the center. Default to 0}

\item{center_y}{numeric the \code{y} coordinate of the center. Default to 0}

\item{point_x}{numeric the \code{x} coordinate of the point}

\item{point_y}{numeric the \code{y} coordinate of the point}
}
\value{
numeric the euclidean distance
}
\description{
Calculate the euclidean distance between two points
}
