% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_emotion_names}
\alias{clean_emotion_names}
\title{clean_emotion_names}
\usage{
clean_emotion_names(data, col)
}
\arguments{
\item{data}{the target dataframe}

\item{col}{name for the target column for the emotions}
}
\value{
the dataframe with the modified column
}
\description{
clean the emotion names for tables and figures (use lazy evaluation)
}
