% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{correct_angle}
\alias{correct_angle}
\title{correct_angle}
\usage{
correct_angle(angle, x, y)
}
\arguments{
\item{angle}{numeric the angle in degrees}

\item{x}{numeric the \code{x} coordinate of the point}

\item{y}{numeric the \code{y} coordinate of the point}
}
\value{
numeric vector of angles
}
\description{
adjust the angle according to quadrants for a 0-centered circle
}
