% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_info_tab}
\alias{create_info_tab}
\title{create_info_tab}
\usage{
create_info_tab(fit_info)
}
\arguments{
\item{fit_info}{the dataframe with fitting information using \code{get_model_info()}}
}
\value{
a dataframe
}
\description{
create_info_tab
}
