% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{deg_to_rad}
\alias{deg_to_rad}
\title{deg_to_rad}
\usage{
deg_to_rad(deg)
}
\arguments{
\item{deg}{numeric the angle in degree}
}
\value{
numeric the angle in radians
}
\description{
convert from radians to degree
}
