% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_all_packages}
\alias{get_all_packages}
\title{get_all_packages}
\usage{
get_all_packages()
}
\value{
a character vector
}
\description{
return all unique packages used in the project. Is a wrapper of \code{renv::dependencies()}
}
