% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_captions}
\alias{get_captions}
\title{get_captions}
\usage{
get_captions(file)
}
\arguments{
\item{file}{}
}
\value{
a named character vector with captions and names
}
\description{
return all captions from a \code{txt} file. All captions need to be on a new line starting with \code{#} and the caption label.
}
