% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_chr_formula}
\alias{get_chr_formula}
\title{get_chr_formula}
\usage{
get_chr_formula(fit_info, collapse = TRUE)
}
\arguments{
\item{fit_info}{the dataframe with fitting information using \code{get_model_info()}}

\item{collapse}{logical that indicate if multiple formulas need to be collapsed into a single character vector.}
}
\value{
a character vector
}
\description{
get_chr_formula
}
