% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_model_info}
\alias{get_model_info}
\title{get_model_info}
\usage{
get_model_info(fit, info = c("formula", "version", "backend"))
}
\arguments{
\item{fit}{an object fitted with \code{brms}}

\item{info}{a character vector with information to be included}
}
\value{
a dataframe
}
\description{
return model fitting information in a dataframe
}
