% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{kappa_to_var}
\alias{kappa_to_var}
\title{Title}
\usage{
kappa_to_var(kappa)
}
\arguments{
\item{kappa}{the kappa value}
}
\value{
the circular variance
}
\description{
converting from \code{k} to the \italic{circula variance}. See https://mathworld.wolfram.com/vonMisesDistribution.html
}
