% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{level_int_position}
\alias{level_int_position}
\title{level_int_position}
\usage{
level_int_position(level_int)
}
\arguments{
\item{level_int}{a numeric vector with distances from the GEW center}
}
\value{
a character vector with the corresponding intensity
}
\description{
assing the intensity label according to the distance from the center of the GEW
}
