% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_named_list}
\alias{make_named_list}
\title{make_named_list}
\usage{
make_named_list(...)
}
\arguments{
\item{...}{a series of objects separated with comma}
}
\value{
a named list with objects and objects' names
}
\description{
make_named_list
}
