% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rad_to_deg}
\alias{rad_to_deg}
\title{rad_to_deg}
\usage{
rad_to_deg(rad)
}
\arguments{
\item{rad}{numeric the angle in radians}
}
\value{
numeric the angle in degree
}
\description{
convert from radians to degree
}
