% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{run_script}
\alias{run_script}
\title{run_script}
\usage{
run_script(file)
}
\arguments{
\item{file}{a character indicating the file path and name}
}
\description{
source a script and return a nice message
}
