% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{seg_position}
\alias{seg_position}
\title{seg_position}
\usage{
seg_position(seg_ang)
}
\arguments{
\item{seg_ang}{a numeric vector with the response angles in degrees}
}
\value{
a character vector with the corresponding emotion label
}
\description{
assigning the emotion label according to the response angle
}
