% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{success_step}
\alias{success_step}
\title{success_step}
\usage{
success_step(x, msg = "completed!")
}
\arguments{
\item{x}{a object}

\item{msg}{the message to be displayed}
}
\description{
return a nice message using the object name. Useful for tracking complex steps.
}
