% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tidy_brm_from_fit}
\alias{tidy_brm_from_fit}
\title{tidy_brm_from_fit}
\usage{
tidy_brm_from_fit(fit)
}
\arguments{
\item{fit}{an object fitted with \code{brms}}
}
\value{
a tibble with all model information
}
\description{
tidy_brm_from_fit
}
