% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tidy_priors}
\alias{tidy_priors}
\title{tidy_priors}
\usage{
tidy_priors(fit)
}
\arguments{
\item{fit}{an object fitted with \code{brms}}
}
\value{
a tibble with all priors information
}
\description{
tidy_priors
}
